requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile requirements.in
#
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.13.1
    # via
    #   fsspec
    #   litellm
aiosignal==1.4.0
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via
    #   httpx
    #   openai
attrs==25.4.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
certifi==2025.10.5
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.4
    # via requests
click==8.1.8
    # via litellm
datasets==4.2.0
    # via -r requirements.in
dill==0.4.0
    # via
    #   datasets
    #   multiprocess
distro==1.9.0
    # via openai
fastuuid==0.13.5
    # via litellm
filelock==3.19.1
    # via
    #   datasets
    #   huggingface-hub
frozenlist==1.8.0
    # via
    #   aiohttp
    #   aiosignal
fsspec[http]==2025.9.0
    # via
    #   datasets
    #   huggingface-hub
h11==0.16.0
    # via httpcore
hf-xet==1.1.10
    # via huggingface-hub
httpcore==1.0.9
    # via httpx
httpx==0.28.1
    # via
    #   datasets
    #   litellm
    #   openai
huggingface-hub==0.35.3
    # via
    #   datasets
    #   tokenizers
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
importlib-metadata==8.7.0
    # via litellm
jinja2==3.1.6
    # via litellm
jiter==0.11.1
    # via openai
jsonschema==4.25.1
    # via litellm
jsonschema-specifications==2025.9.1
    # via jsonschema
litellm==1.78.3
    # via -r requirements.in
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.3
    # via jinja2
mdurl==0.1.2
    # via markdown-it-py
multidict==6.7.0
    # via
    #   aiohttp
    #   yarl
multiprocess==0.70.16
    # via datasets
numpy==2.3.4
    # via
    #   datasets
    #   pandas
openai==2.5.0
    # via litellm
packaging==25.0
    # via
    #   datasets
    #   huggingface-hub
pandas==2.3.3
    # via datasets
propcache==0.4.1
    # via
    #   aiohttp
    #   yarl
pyarrow==21.0.0
    # via datasets
pydantic==2.12.3
    # via
    #   -r requirements.in
    #   litellm
    #   openai
pydantic-core==2.41.4
    # via pydantic
pygments==2.19.2
    # via rich
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.1.1
    # via
    #   -r requirements.in
    #   litellm
pytz==2025.2
    # via pandas
pyyaml==6.0.3
    # via
    #   -r requirements.in
    #   datasets
    #   huggingface-hub
referencing==0.37.0
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2025.9.18
    # via tiktoken
requests==2.32.5
    # via
    #   datasets
    #   huggingface-hub
    #   tiktoken
rich==14.2.0
    # via -r requirements.in
rpds-py==0.27.1
    # via
    #   jsonschema
    #   referencing
six==1.17.0
    # via python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   openai
tenacity==9.1.2
    # via -r requirements.in
tiktoken==0.12.0
    # via litellm
tokenizers==0.22.1
    # via litellm
tqdm==4.67.1
    # via
    #   datasets
    #   huggingface-hub
    #   openai
typing-extensions==4.15.0
    # via
    #   aiosignal
    #   anyio
    #   huggingface-hub
    #   openai
    #   pydantic
    #   pydantic-core
    #   referencing
    #   typing-inspection
typing-inspection==0.4.2
    # via pydantic
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via requests
xxhash==3.6.0
    # via datasets
yarl==1.22.0
    # via aiohttp
zipp==3.23.0
    # via importlib-metadata
